feat: add share wallet ui
