+++ packages/maskbook/src/_locales/en/messages.json
@@ -452,5 +452,6 @@
     "plugin_ito_claim_only_once_remind": "Each wallet can only participate once.",
     "wallet_transfer_to_address": "To Address",
     "wallet_transfer_send": "Send",
-    "wallet_transfer_balance": "Balance"
+    "wallet_transfer_balance": "Balance",
+    "wallet_transfer_title": "Transfer"
 }
